Fix -Wshorten64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@39571 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tools/cartconv/cartconv.c
@@ -1056,7 +1056,7 @@ static int checkflag(char *flg, char *arg)
                 checkarg(arg);
                 if (load_address == 0) {
                     /* load_address = atoi(arg); */
-                    load_address = strtoul(arg, NULL, 0);
+                    load_address = (int)strtoul(arg, NULL, 0);
                 } else {
                     usage();
                 }
